refactor: [WIP] Split file about patterns
